@@ -3,5 +3,5 @@ name = "credis-python"
 version = "0.1.0"
 description = "Redis controller client for Sentinel Setup"
 readme = "README.md"
-requires-python = ">=3.9"
-dependencies = ["redis>=5.2.1"]
+requires-python = ">=3.10"
+dependencies = ["dill>=0.4.0", "redis>=6.4.0"]
